Trapped: Handle large heatmaps.
If we have a heatmap with entries that have been "rounded" down,
we pass this information to Javascript, and use it to 1) mark
the headers, and 2) use an orange background

+++ Src/C/heatmap.h
@@ -4,8 +4,8 @@
 
 # include "trapped.h"
 
-# define HEATMAP_ROW_SIZE 10
-# define HEATMAP_COL_SIZE 10
+# define HEATMAP_ROW_SIZE ((short unsigned) 10)
+# define HEATMAP_COL_SIZE ((short unsigned) 10)
 # define HEATMAP_NONE      0
 # define HEATMAP_ABS      (1 << 0)
 # define HEATMAP_PERC     (1 << 1)
